# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    ApproveCampaignCriterionRequest:
      type: object
      properties:
        note:
          type: string
...
    Archetype:
      type: string
      enum:
      - EPIC
      - TICKET
      - FEATURE
      - TASK
      - CAMPAIGN
...
    CampaignMemberResponse:
      type: object
      properties:
        member:
          $ref: "#/components/schemas/CampaignMemberDto"
...
    CampaignResponse:
      type: object
      properties:
        campaign:
          $ref: "#/components/schemas/CampaignDto"
...
    EntityTransition:
      type: object
      properties:
        archetype:
          $ref: "#/components/schemas/Archetype"
        membership:
          $ref: "#/components/schemas/Membership"
        title:
          type: string
        description:
          type: string
        status:
          type: string
        ticketType:
          $ref: "#/components/schemas/TicketType"
        impetus:
          type: string
        assignee:
          type: string
        supersededBy:
          type: string
        repositoryId:
          type: string
        implementedAt:
          $ref: "#/components/schemas/Instant"
        dependsOn:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
...
    MoveCampaignMemberRequest:
      type: object
      required:
      - position
      properties:
        position:
          type: integer
          format: int32
...
    SetCampaignMemberConditionRequest:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/ConditionGroup"
...
    WorkDeleted:
      type: object
      description: The entity and its subtree are gone.
      properties:
        success:
          type: boolean
...
    WorkMemberAdd:
      type: object
      required:
      - entityId
      description: An entity to gather into the campaign.
      properties:
        entityId:
          type: string
          pattern: \S
          description: "The entity to gather, by qualified id (<projectSlug>-<n>)\
            \ or UUID."
        position:
          type: integer
          format: int32
          description: "Where it goes, 0-based; omitted appends."
        inFlight:
          type: boolean
          description: Whether it joins already running; omitted lets the service
            decide from its status and its workspace.
    WorkMemberAnswer:
      type: object
      description: One membership of a campaign.
      properties:
        member:
          $ref: "#/components/schemas/CampaignMemberDto"
...
    WorkMemberList:
      type: object
      description: "A campaign's members, in campaign order."
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/CampaignMemberDto"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /projects/api/campaigns/{id}/members/{membershipId}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Remove Member
      tags:
      - Campaign Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /projects/api/campaigns/{id}/members/{membershipId}/condition:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCampaignMemberConditionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignMemberResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Set Condition
      tags:
      - Campaign Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /projects/api/campaigns/{id}/members/{membershipId}/criteria/{criterionId}/approve:
    post:
      parameters:
      - name: criterionId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveCampaignCriterionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignMemberResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Approve
      tags:
      - Campaign Controller
      security:
      - SecurityScheme:
        - qits:admin
  /projects/api/campaigns/{id}/members/{membershipId}/position:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveCampaignMemberRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Move Member
      tags:
      - Campaign Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
...
  /projects/api/work/{qualifiedId}:
    put:
      summary: State a work entity in full
      description: "The full edit of one entity: the body is the entity's whole state\
        \ afterwards — what it is (archetype), where it hangs (membership), every\
        \ property — and an absent property is cleared, not left alone. It is the\
        \ one-entry form of POST /projects/api/work/transition and obeys every rule\
        \ of it, so it is also how a single entity is reshaped, reparented or superseded.\
        \ Every entity id in the body (membership.parent, supersededBy, dependsOn)\
        \ is a qualified id or a UUID. Answers the entity as written, in the merged\
        \ shape."
      operationId: putWork
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityTransition"
        required: true
      responses:
        "200":
          description: The entity as written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionedEntity"
        "400":
          description: "Every violation of the stated state, in one message"
        "403":
          description: An agent writing outside its own project; nothing is written
        "404":
          description: No entity with this id
        "409":
          description: "A status this door may not state (a change to or from READY_FOR_DEV,\
            \ a jump into started work, a gated move — those go through POST /work/{qualifiedId}/status),\
            \ or a changed acceptance-criteria list from READY_FOR_DEV on"
        "401":
          description: Not Authorized
      tags:
      - Work Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
    patch:
      summary: Edit a work entity's fields (JSON merge patch)
      description: "Partial update of one entity of any archetype. An absent property\
        \ is left unchanged and an explicit null clears it. The status is never written\
        \ here — a lifecycle move goes through POST /projects/api/work/{qualifiedId}/status,\
        \ and a reshape, reparent or supersede through PUT /projects/api/work/{qualifiedId}\
        \ or POST /projects/api/work/transition. dependsOn is a qualified id or a\
        \ UUID. An epic's, feature's or task's scope follows the epic's freeze: scope\
        \ edits need the epic REPORTED, the markers need it READY_FOR_DEV or IMPLEMENTING,\
        \ and a marker moves the item's status with it. An epic's or a ticket's acceptance\
        \ criteria are frozen from READY_FOR_DEV on, where a changed list is refused\
        \ and the same list restated passes. Answers the entity in the merged shape."
      operationId: patchWork
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/merge-patch+json:
            schema:
              $ref: "#/components/schemas/WorkPatch"
          application/json:
            schema:
              $ref: "#/components/schemas/WorkPatch"
        required: true
      responses:
        "200":
          description: The entity as it stands after the edit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionedEntity"
        "400":
          description: "Every complaint about the body in one message: not an object,\
            \ empty, a property that is a move or server-owned or unknown, one the\
            \ archetype has no slot for, a null where the property cannot be cleared,\
            \ or a value of the wrong shape"
        "403":
          description: An agent writing an entity outside its own project; nothing
            is written
        "404":
          description: No entity with this id
        "409":
          description: "The owning epic's status freezes what the patch touches, or\
            \ the acceptance criteria are changed from READY_FOR_DEV on"
        "401":
          description: Not Authorized
      tags:
      - Work Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
        - qits:system
    get:
      summary: Read one work entity of any archetype
      description: "The entity in the merged shape, with its qualified id, its parent\
        \ and position, and whether it is blocked. The path names it by qualified\
        \ id (<projectSlug>-<n>) or UUID."
      operationId: getWork
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionedEntity"
        "404":
          description: No entity with this id
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
        - qits:system
    delete:
      summary: Delete a work entity and its subtree
      description: "Removes the entity, its descendants and their threads, each audited;\
        \ the audit log outlives them. An epic's or a ticket's delete is qits:admin\
        \ alone; a feature's or a task's admits an agent bound to its project and\
        \ obeys its epic's freeze. A campaign is not deleted but dropped. The path\
        \ names the entity by qualified id (<projectSlug>-<n>) or UUID."
      operationId: deleteWork
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The entity is gone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkDeleted"
        "403":
          description: "An agent deleting an epic or a ticket, or outside its own\
            \ project"
        "404":
          description: No entity with this id
        "409":
          description: "A campaign, or a feature or a task whose epic is no longer\
            \ REPORTED"
        "401":
          description: Not Authorized
      tags:
      - Work Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
...
  /projects/api/work/{qualifiedId}/members:
    get:
      summary: A campaign's members
      description: "Every membership in campaign order: the entity it gathers, its\
        \ run record and its condition. The path names the campaign by qualified id\
        \ (<projectSlug>-<n>) or UUID."
      operationId: listWorkMembers
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkMemberList"
        "404":
          description: No campaign with this id
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Work Members Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
    post:
      summary: Gather an entity into a campaign
      description: "Adds a membership — at position, or at the end — for an entity\
        \ named by qualified id or UUID. A bound agent adds only in its own project."
      operationId: addWorkMember
      parameters:
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkMemberAdd"
        required: true
      responses:
        "201":
          description: The membership as written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkMemberAnswer"
        "400":
          description: No entityId
        "403":
          description: An agent outside its own project
        "404":
          description: "No campaign, or no entity, with this id"
        "409":
          description: "The entity cannot join: already a member, a feature or a task,\
            \ a closed campaign"
        "401":
          description: Not Authorized
      tags:
      - Work Members Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /projects/api/work/{qualifiedId}/members/{membershipId}:
    delete:
      summary: Remove a membership from its campaign
      description: "Refused once the member was claimed, and while another member's\
        \ criterion targets it. Answers the members that remain."
      operationId: removeWorkMember
      parameters:
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The members that remain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkMemberList"
        "403":
          description: An agent outside its own project
        "404":
          description: "No campaign, or no such membership on it"
        "409":
          description: "Claimed, or targeted by another member"
        "401":
          description: Not Authorized
      tags:
      - Work Members Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /projects/api/work/{qualifiedId}/members/{membershipId}/condition:
    put:
      summary: Replace a membership's condition
      description: "PUT semantics: the whole condition, groups OR'd and criteria in\
        \ a group AND'd; an empty list waits on nothing. A criterion restating an\
        \ existing id keeps its latch. An ENTITY_STATUS predicate's entityId takes\
        \ a qualified id or a UUID."
      operationId: setWorkMemberCondition
      parameters:
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkMemberCondition"
        required: true
      responses:
        "200":
          description: The member with its condition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkMemberAnswer"
        "400":
          description: A criterion the service cannot read
        "403":
          description: An agent outside its own project
        "404":
          description: "No campaign, or no such membership on it"
        "401":
          description: Not Authorized
      tags:
      - Work Members Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
  /projects/api/work/{qualifiedId}/members/{membershipId}/criteria/{criterionId}/approve:
    post:
      summary: A person's yes on a member's APPROVAL criterion
      description: "qits:admin, and then a person this service verified itself; the\
        \ name recorded is the proof's. Refused when the criterion is not APPROVAL,\
        \ is already satisfied, or the campaign is DONE or DROPPED."
      operationId: approveWorkMemberCriterion
      parameters:
      - name: criterionId
        in: path
        required: true
        schema:
          type: string
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkCriterionApproval"
        required: true
      responses:
        "200":
          description: The member with the criterion satisfied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkMemberAnswer"
        "403":
          description: Not a verified person
        "404":
          description: "No campaign, membership or criterion"
        "409":
          description: "Not an APPROVAL, already satisfied, or closed"
        "401":
          description: Not Authorized
        "400":
          description: Bad Request
      tags:
      - Work Members Controller
      security:
      - SecurityScheme:
        - qits:admin
  /projects/api/work/{qualifiedId}/members/{membershipId}/position:
    put:
      summary: Move a membership within its campaign
      description: Answers the members as the move left them.
      operationId: moveWorkMember
      parameters:
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
      - name: qualifiedId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkMemberMove"
        required: true
      responses:
        "200":
          description: "The members, in their new order"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkMemberList"
        "403":
          description: An agent outside its own project
        "404":
          description: "No campaign, or no such membership on it"
        "401":
          description: Not Authorized
        "400":
          description: Bad Request
      tags:
      - Work Members Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:agent
...
info:
  title: qits-projects API
  version: 2026.1006.121753